Split sample browser row tests
